Bypass Prompt for Pre-Approved Tools
## [1mStep 1[22m

In the `createPermissionSocketServer` method, at the beginning of the `socket.on('data', ...)` event handler, add logic to check if the `tool_name` from the incoming message exists in the `this.alwaysAllowedTools` set.


## [1mStep 2[22m

If the `tool_name` is found in the set, send an `approved: true` response directly to the socket and return from the handler to bypass the user prompt. Be sure to add a log message indicating that the tool was automatically approved.