Update State When "Always Allow" is Selected
## [1mStep 1[22m

In `createPermissionSocketServer`, modify the logic that processes the result from the `select` prompt.


## [1mStep 2[22m

If the user's choice is "Always Allow", add the `tool_name` from the permission request to the `this.alwaysAllowedTools` set before sending the approval response. The logic to set `approved` to `true` should already be in place from the first task.