Enhance Storage and Integrate Bash Prefix Prompt
## [1mStep 1[22m

In `src/rmplan/executors/claude_code.ts`, change the `alwaysAllowedTools` class property from `new Set<string>()` to `new Map<string, true | string[]>()`. This will allow storing either a boolean for general tools or an array of string prefixes specifically for the `Bash` tool.


## [1mStep 2[22m

Import the new custom prefix prompt you created in the previous task into `claude_code.ts`.


## [1mStep 3[22m

In the `createPermissionSocketServer` method, locate the logic where `userChoice === 'always_allow'`. Add a condition to check if the `tool_name` is 'Bash'.


## [1mStep 4[22m

If the tool is 'Bash', call your new prefix prompt, passing the `input.command` to it. When the prompt resolves with a selected prefix, add this prefix to the array of allowed prefixes for 'Bash' in the `alwaysAllowedTools` map. If an array for 'Bash' doesn't exist yet, create it.


## [1mStep 5[22m

Update the logic for non-Bash tools. When "Always Allow" is chosen, the behavior should be to set `this.alwaysAllowedTools.set(tool_name, true)`.